WIP: code to edit progress
