--- gui/src/components/CodeToEditCard/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./CodeToEditCard";
