+++ gui/src/redux/slices/editModeState.ts
@@ -2,6 +2,8 @@ import { createSlice, PayloadAction } from "@reduxjs/toolkit";
 import { ApplyState, ApplyStateStatus, CodeToEdit, MessageModes } from "core";
 import { EDIT_MODE_STREAM_ID } from "core/edit/constants";
 export interface EditModeState {
+  // Array because of previous multi-file edit functionality
+  // Keeping array to not break persisted redux for now
   codeToEdit: CodeToEdit[];
   applyState: ApplyState;
   returnCursorToEditorAfterEdit: boolean;
@@ -52,20 +54,9 @@ export const editModeStateSlice = createSlice({
       }>,
     ) => {
       state.returnCursorToEditorAfterEdit = payload.fromEditor;
-      const entries = Array.isArray(payload.codeToEdit)
+      state.codeToEdit = Array.isArray(payload.codeToEdit)
         ? payload.codeToEdit
         : [payload.codeToEdit];
-
-      const newEntries = entries.filter(
-        (entry) =>
-          !state.codeToEdit.some((existingEntry) =>
-            isCodeToEditEqual(existingEntry, entry),
-          ),
-      );
-
-      if (newEntries.length > 0) {
-        state.codeToEdit.push(...newEntries);
-      }
     },
     clearCodeToEdit: (state) => {
       state.codeToEdit = [];
@@ -83,22 +74,3 @@ export const {
   setEditStateApplyState,
 } = editModeStateSlice.actions;
 export default editModeStateSlice.reducer;
-
-function isCodeToEditEqual(a: CodeToEdit, b: CodeToEdit) {
-  if (a.filepath !== b.filepath || a.contents !== b.contents) {
-    return false;
-  }
-
-  if ("range" in a && "range" in b) {
-    const rangeA = a.range;
-    const rangeB = b.range;
-
-    return (
-      rangeA.start.line === rangeB.start.line &&
-      rangeA.end.line === rangeB.end.line
-    );
-  }
-
-  // If neither has a range, they are considered equal in this context
-  return !("range" in a) && !("range" in b);
-}
